+++ client/src/plugins/resource/archive/rtsp_client_archive_delegate.cpp
@@ -163,10 +163,10 @@ void QnRtspClientArchiveDelegate::checkMinTimeFromOtherServer(const QnVirtualCam
         return;
 
     qint64 startTime = otherRtspSession.startTime();
-    if ((quint64)startTime != AV_NOPTS_VALUE && startTime != DATETIME_NOW) 
+    if (startTime != qint64(AV_NOPTS_VALUE) && startTime != DATETIME_NOW)
     {
         QnMutexLocker lock( &m_timeMutex );
-        if (startTime < *result || *result == AV_NOPTS_VALUE)
+        if (startTime < *result || *result == qint64(AV_NOPTS_VALUE))
             *result = startTime;
     }
 }
@@ -181,7 +181,7 @@ void QnRtspClientArchiveDelegate::checkMinTimeFromOtherServer(const QnVirtualCam
 {
     if (!camera) {
         QnMutexLocker lock( &m_timeMutex );
-        m_globalMinArchiveTime = AV_NOPTS_VALUE;
+        m_globalMinArchiveTime = qint64(AV_NOPTS_VALUE);
         return;
     }
 
@@ -189,21 +189,21 @@ void QnRtspClientArchiveDelegate::checkMinTimeFromOtherServer(const QnVirtualCam
     /* Check if no archive available on any server. */
     if (mediaServerList.isEmpty()) {
         QnMutexLocker lock( &m_timeMutex );
-        m_globalMinArchiveTime = AV_NOPTS_VALUE;
+        m_globalMinArchiveTime = qint64(AV_NOPTS_VALUE);
         return;
     }
 
     QList<ArchiveTimeCheckInfo> checkList;
-    qint64 currentMinTime  = AV_NOPTS_VALUE;
-    qint64 otherMinTime  = AV_NOPTS_VALUE;
+    qint64 currentMinTime  = qint64(AV_NOPTS_VALUE);
+    qint64 otherMinTime  = qint64(AV_NOPTS_VALUE);
     for (const auto &server: mediaServerList)
         checkList << ArchiveTimeCheckInfo(camera, server, this, server == m_server ? &currentMinTime : &otherMinTime);
     QtConcurrent::blockingFilter(checkList, checkGlobalMinTime);
     QnMutexLocker lock( &m_timeMutex );
-    if ((otherMinTime != AV_NOPTS_VALUE) && (currentMinTime == AV_NOPTS_VALUE || otherMinTime < currentMinTime))
+    if ((otherMinTime != qint64(AV_NOPTS_VALUE)) && (currentMinTime == qint64(AV_NOPTS_VALUE) || otherMinTime < currentMinTime))
         m_globalMinArchiveTime = otherMinTime;
     else
-        m_globalMinArchiveTime = AV_NOPTS_VALUE;
+        m_globalMinArchiveTime = qint64(AV_NOPTS_VALUE);
 }
 
 QnMediaServerResourcePtr QnRtspClientArchiveDelegate::getServerOnTime(qint64 timeUsec) {
@@ -253,7 +253,7 @@ bool QnRtspClientArchiveDelegate::openInternal() {
         m_server = getServerOnTime(m_position); // try to update server
         if (m_server == 0 || m_server->getStatus() == Qn::Offline) 
         {
-            if (m_isMultiserverAllowed && m_globalMinArchiveTime == AV_NOPTS_VALUE)
+            if (m_isMultiserverAllowed && m_globalMinArchiveTime == qint64(AV_NOPTS_VALUE))
                 checkMinTimeFromOtherServer(m_camera);
             return false;
         }
@@ -350,15 +350,15 @@ void QnRtspClientArchiveDelegate::lockTime(qint64 value)
 void QnRtspClientArchiveDelegate::unlockTime()
 {
     QnMutexLocker lock( &m_timeMutex );
-    m_lockedTime = AV_NOPTS_VALUE;
+    m_lockedTime = qint64(AV_NOPTS_VALUE);
 }
 
 qint64 QnRtspClientArchiveDelegate::startTime()
 {
     QnMutexLocker lock( &m_timeMutex );
-    if(m_lockedTime != AV_NOPTS_VALUE)
+    if(m_lockedTime != qint64(AV_NOPTS_VALUE))
         return m_lockedTime;
-    qint64 result = m_globalMinArchiveTime != AV_NOPTS_VALUE ? m_globalMinArchiveTime : m_rtspSession.startTime();
+    qint64 result = m_globalMinArchiveTime != qint64(AV_NOPTS_VALUE) ? m_globalMinArchiveTime : m_rtspSession.startTime();
 
     if (result == DATETIME_NOW || result <= qnSyncTime->currentMSecsSinceEpoch()*1000)
         return result;
@@ -496,7 +496,7 @@ QnAbstractMediaDataPtr QnRtspClientArchiveDelegate::getNextDataInternal()
             rtpChannelNum = m_rtpData->getMediaSocket()->getLocalAddress().port;
         }
         const QString format = m_rtspSession.getTrackFormatByRtpChannelNum(rtpChannelNum).toLower();
-        qint64 parserPosition = AV_NOPTS_VALUE;
+        qint64 parserPosition = qint64(AV_NOPTS_VALUE);
         if (format.isEmpty()) {
             //qWarning() << Q_FUNC_INFO << __LINE__ << "RTP track" << rtpChannelNum << "not found";
         }
@@ -513,7 +513,7 @@ QnAbstractMediaDataPtr QnRtspClientArchiveDelegate::getNextDataInternal()
 
         if (result && m_sendedCSec != result->opaque)
             result.reset(); // ignore old archive data
-        if (result && parserPosition != AV_NOPTS_VALUE)
+        if (result && parserPosition != qint64(AV_NOPTS_VALUE))
             m_position = parserPosition;
         /*
         if (result && m_waitBOF)
